requirements.txt
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pytest==8.3.5
